+++ LdapForNet/SearchRequestHandler.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.InteropServices;
 using LdapForNet.Native;
 using LdapForNet.Utils;
@@ -8,14 +7,6 @@ using static LdapForNet.Native.Native;
 
 namespace LdapForNet
 {
-    internal interface IRequestHandler
-    {
-        int SendRequest(SafeHandle handle, DirectoryRequest request, ref int messageId);
-
-        LdapResultCompleteStatus Handle(SafeHandle handle, LdapResultType resType, IntPtr msg,
-            out DirectoryResponse response);
-    }
-
     internal class SearchRequestHandler : IRequestHandler
     {
         private readonly SearchResponse _response = new SearchResponse();
@@ -33,7 +24,7 @@ namespace LdapForNet
                     IntPtr.Zero,
                     IntPtr.Zero,
                     IntPtr.Zero,
-                    (int) LdapSizeLimit.LDAP_NO_LIMIT,
+                    500,
                     ref messageId);
             }
 
@@ -107,237 +98,5 @@ namespace LdapForNet
         }
 
 
-    }
-
-    internal class AddRequestHandler : IRequestHandler
-    {
-        public int SendRequest(SafeHandle handle, DirectoryRequest request, ref int messageId)
-        {
-            if (request is AddRequest addRequest)
-            {
-                var entry = addRequest.LdapEntry;
-                if (string.IsNullOrWhiteSpace(entry.Dn))
-                {
-                    throw new ArgumentNullException(nameof(entry.Dn));
-                }
-
-                if (entry.Attributes == null)
-                {
-                    entry.Attributes = new Dictionary<string, List<string>>();
-                }
-
-                var attrs = entry.Attributes.Select(ToLdapMod).ToList();
-
-                var ptr = Marshal.AllocHGlobal(IntPtr.Size*(attrs.Count+1)); // alloc memory for list with last element null
-                MarshalUtils.StructureArrayToPtr(attrs,ptr, true);
-
-                return LdapNative.Instance.ldap_add_ext(handle,
-                    addRequest.LdapEntry.Dn,
-                    ptr,                
-                    IntPtr.Zero, 
-                    IntPtr.Zero ,
-                    ref messageId
-                );    
-            }
-
-            return 0;
-
-        }
-
-        public LdapResultCompleteStatus Handle(SafeHandle handle, LdapResultType resType, IntPtr msg, out DirectoryResponse response)
-        {
-            response = default;
-            switch (resType)
-            {
-                case LdapResultType.LDAP_RES_ADD:
-                    response = new AddResponse();
-                    return LdapResultCompleteStatus.Complete;
-                default:
-                    return LdapResultCompleteStatus.Unknown;
-            }
-        }
-        
-        private static LDAPMod ToLdapMod(KeyValuePair<string, List<string>> attribute)
-        {
-            return ToLdapMod(new LdapModifyAttribute
-            {
-                Type = attribute.Key,
-                LdapModOperation = LdapModOperation.LDAP_MOD_ADD,
-                Values = attribute.Value
-            });
-        }
-        
-        private static LDAPMod ToLdapMod(LdapModifyAttribute attribute)
-        {
-            var modValue = GetModValue(attribute.Values);
-            var modValuePtr = Marshal.AllocHGlobal(IntPtr.Size * (modValue.Count));
-            MarshalUtils.StringArrayToPtr(modValue, modValuePtr);
-            return new LDAPMod
-            {
-                mod_op = (int) attribute.LdapModOperation,
-                mod_type = attribute.Type,
-                mod_vals_u = new LDAPMod.mod_vals
-                {
-                    modv_strvals = modValuePtr,
-                },
-                mod_next = IntPtr.Zero
-            };
-        }
-        
-        private static List<string> GetModValue(List<string> values)
-        {
-            var res = values??new List<string>();
-            res.Add(null);
-            return res;
-        }
-
-
-    }
-
-    internal class ModifyRequestHandler : IRequestHandler
-    {
-        public int SendRequest(SafeHandle handle, DirectoryRequest request, ref int messageId)
-        {
-            if (request is ModifyRequest modifyRequest)
-            {
-                var entry = modifyRequest.LdapEntry;
-                if (string.IsNullOrWhiteSpace(entry.Dn))
-                {
-                    throw new ArgumentNullException(nameof(entry.Dn));
-                }
-            
-                if (entry.Attributes == null)
-                {
-                    entry.Attributes = new List<LdapModifyAttribute>();
-                }
-            
-                var attrs = entry.Attributes.Select(ToLdapMod).ToList();
-            
-                var ptr = Marshal.AllocHGlobal(IntPtr.Size*(attrs.Count+1)); // alloc memory for list with last element null
-                MarshalUtils.StructureArrayToPtr(attrs,ptr, true);
-                
-                return LdapNative.Instance.ldap_modify_ext(handle,
-                    entry.Dn,
-                    ptr,                
-                    IntPtr.Zero, 
-                    IntPtr.Zero ,
-                    ref messageId
-                );    
-            }
-
-            return 0;
-        }
-
-        public LdapResultCompleteStatus Handle(SafeHandle handle, LdapResultType resType, IntPtr msg, out DirectoryResponse response)
-        {
-            response = default;
-            switch (resType)
-            {
-                case LdapResultType.LDAP_RES_MODIFY:
-                    response = new ModifyResponse();
-                    return LdapResultCompleteStatus.Complete;
-                default:
-                    return LdapResultCompleteStatus.Unknown;
-            }
-        }
-        
-        private static LDAPMod ToLdapMod(LdapModifyAttribute attribute)
-        {
-            var modValue = GetModValue(attribute.Values);
-            var modValuePtr = Marshal.AllocHGlobal(IntPtr.Size * (modValue.Count));
-            MarshalUtils.StringArrayToPtr(modValue, modValuePtr);
-            return new LDAPMod
-            {
-                mod_op = (int) attribute.LdapModOperation,
-                mod_type = attribute.Type,
-                mod_vals_u = new LDAPMod.mod_vals
-                {
-                    modv_strvals = modValuePtr,
-                },
-                mod_next = IntPtr.Zero
-            };
-        }
-        
-        private static List<string> GetModValue(List<string> values)
-        {
-            var res = values??new List<string>();
-            res.Add(null);
-            return res;
-        }
-    }
-
-    internal class DeleteRequestHandler : IRequestHandler
-    {
-        public int SendRequest(SafeHandle handle, DirectoryRequest request, ref int messageId)
-        {
-            if (request is DeleteRequest deleteRequest)
-            {
-                var dn = deleteRequest.DistinguishedName;
-                if (string.IsNullOrWhiteSpace(dn))
-                {
-                    throw new ArgumentNullException(nameof(dn));
-                }
-                return LdapNative.Instance.ldap_delete_ext(handle,
-                    dn,
-                    IntPtr.Zero, 
-                    IntPtr.Zero ,    
-                    ref messageId
-                );  
-            }
-
-            return 0;
-        }
-
-        public LdapResultCompleteStatus Handle(SafeHandle handle, LdapResultType resType, IntPtr msg, out DirectoryResponse response)
-        {
-            response = default;
-            switch (resType)
-            {
-                case LdapResultType.LDAP_RES_DELETE:
-                    response = new DeleteResponse();
-                    return LdapResultCompleteStatus.Complete;
-                default:
-                    return LdapResultCompleteStatus.Unknown;
-            }
-        }
-    }
-    
-    internal class ModifyDnRequestHandler: IRequestHandler
-    {
-        public int SendRequest(SafeHandle handle, DirectoryRequest request, ref int messageId)
-        {
-            if (request is ModifyDNRequest modifyDnRequest)
-            {
-                var dn = modifyDnRequest.DistinguishedName;
-                if (string.IsNullOrWhiteSpace(dn))
-                {
-                    throw new ArgumentNullException(nameof(dn));
-                }
-                return LdapNative.Instance.ldap_rename(handle,
-                    dn,
-                    modifyDnRequest.NewName,
-                    modifyDnRequest.NewParentDistinguishedName ,    
-                    modifyDnRequest.DeleteOldRdn?1:0,
-                    IntPtr.Zero, 
-                    IntPtr.Zero, 
-                    ref messageId
-                );  
-            }
-
-            return 0;
-        }
-
-        public LdapResultCompleteStatus Handle(SafeHandle handle, LdapResultType resType, IntPtr msg, out DirectoryResponse response)
-        {
-            response = default;
-            switch (resType)
-            {
-                case LdapResultType.LDAP_RES_MODDN:
-                    response = new ModifyDNResponse();
-                    return LdapResultCompleteStatus.Complete;
-                default:
-                    return LdapResultCompleteStatus.Unknown;
-            }
-        }
     }
 }
